+++ 0x05-pointers_arrays_strings/holberton.h
@@ -7,7 +7,7 @@ void swap_int(int *a, int *b);
 int _strlen(char *s);
 void _puts(char *str);
 void print_rev(char *s);
-int incase(char *s,int num);
+int incase(char *s,long int num);
 int _atoi(char *s);
 
 #endif
